backend/requirements.txt
# Kahflane Backend Dependencies

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
starlette>=0.35.0

# Database
sqlmodel>=0.0.14
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0,<5.0.0
pyjwt>=2.8.0

# Validation & Settings
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Redis & Job Queue
redis>=5.0.0
dramatiq[redis]>=1.15.0

# Vector Database
qdrant-client>=1.7.0

# AI & ML
google-genai>=1.0.0
elevenlabs>=1.0.0
# crewai>=0.1.0  # Uncomment when needed

# Document Processing
docling>=2.0.0
docling-core[chunking]>=2.0.0
python-docx>=1.1.0
PyPDF2>=3.0.0
python-multipart>=0.0.6

# Storage
boto3>=1.34.0  # For S3 storage
aiofiles>=23.2.0

# HTTP Client
httpx>=0.26.0

# Utilities
python-dotenv>=1.0.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
black>=24.1.0
ruff>=0.1.0
mypy>=1.8.0
